Readding import which was removed. Removing redundant lines. Preventing remove_old_html from removing the template.

+++ flv.py
@@ -1,23 +1,25 @@
 #!/usr/bin/env python3
+from pathlib import Path
 
 
 def generate_html():
     with open("template.html") as f:
         template = f.read()
     p = Path(".")
-    swfs = list(p.glob("**/*.swf"))
-    for swf in swfs:
+    for swf in p.glob("**/*.swf"):
         content = template.replace("template", swf.stem)
         html = swf.with_suffix(".html")
-        with open(html, "w") as f:
-            f.write(content)
+        html.write_text(content)
 
 
 def remove_old_html():
     p = Path(".")
     for html in p.glob("**/*.html"):
+        if html.name == "template.html":
+            continue
         html.unlink()
+
 
 if __name__ == "__main__":
     remove_old_html()
-    #generate_html()
+    generate_html()
